basically did mvp

+++ app/src/main/java/com/example/daily/data/local/entity/TopicSpecDto.kt
@@ -1,6 +1,6 @@
 package com.example.daily.data.local.entity
 
 data class TopicSpecDto(
-    val color: Int,
+    val color: Long,
     val description: String
 )
