+++ app/src/main/java/com/example/daily/model/TopicSpec.kt
@@ -1,6 +1,6 @@
 package com.example.daily.model
 
 data class TopicSpec(
-    val color: Int,
+    val color: Long,
     val description: String
 )
